git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/bulkComms@20446 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ runtime/include/chplcgfns.h
@@ -9,7 +9,7 @@
 /* defined in chpl__header.h: */
 
 extern int64_t numThreadsPerLocale;
-extern int64_t chpl__maxThreadsPerLocale;
+extern int32_t chpl__maxThreadsPerLocale;
 extern int64_t callStackSize;
 
 /* generated */
